+++ data/examples/declaration/value/function/do/recursive-do-mdo-out.hs
@@ -1,12 +1,11 @@
 {-# LANGUAGE RecursiveDo #-}
 
-baz =
-  mdo
-    bar a
-    a <- foo
-    b <-
-      bar
-        1
-        2
-        3
-    return (a + b)
+baz = mdo
+  bar a
+  a <- foo
+  b <-
+    bar
+      1
+      2
+      3
+  return (a + b)